consensus_encoding: move encoders inline tests to integration
